+++ nifi-nar-bundles/nifi-kudu-bundle/nifi-kudu-processors/src/test/java/org/apache/nifi/processors/kudu/MockPutKudu.java
@@ -17,61 +17,64 @@
 
 package org.apache.nifi.processors.kudu;
 
+import org.apache.kudu.client.Insert;
 import org.apache.kudu.client.KuduClient;
 import org.apache.kudu.client.KuduSession;
 import org.apache.kudu.client.KuduTable;
-import org.apache.kudu.client.KuduException;
-import org.apache.kudu.client.Insert;
 import org.apache.kudu.client.Upsert;
-
 import org.apache.nifi.serialization.record.Record;
 
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 
+import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class MockPutKudu extends PutKudu {
-  private KuduSession session;
-  private LinkedList<Insert> insertQueue;
+    private KuduSession session;
+    private LinkedList<Insert> insertQueue;
 
-  public MockPutKudu() {
-    this(mock(KuduSession.class));
-  }
+    public MockPutKudu() {
+        this(mock(KuduSession.class));
+    }
 
-  public MockPutKudu(KuduSession session) {
-    this.session = session;
-    this.insertQueue = new LinkedList<>();
-  }
+    public MockPutKudu(KuduSession session) {
+        this.session = session;
+        this.insertQueue = new LinkedList<>();
+    }
 
-  public void queue(Insert... operations) {
-    insertQueue.addAll(Arrays.asList(operations));
-  }
+    public void queue(Insert... operations) {
+        insertQueue.addAll(Arrays.asList(operations));
+    }
 
-  @Override
-  protected Insert insertRecordToKudu(KuduTable kuduTable, Record record, List<String> fieldNames) {
-    Insert insert = insertQueue.poll();
-    return insert != null ? insert : mock(Insert.class);
-  }
+    @Override
+    protected Insert insertRecordToKudu(KuduTable kuduTable, Record record, List<String> fieldNames) {
+        Insert insert = insertQueue.poll();
+        return insert != null ? insert : mock(Insert.class);
+    }
 
-  @Override
-  protected Upsert upsertRecordToKudu(KuduTable kuduTable, Record record, List<String> fieldNames) {
-    return mock(Upsert.class);
-  }
+    @Override
+    protected Upsert upsertRecordToKudu(KuduTable kuduTable, Record record, List<String> fieldNames) {
+        return mock(Upsert.class);
+    }
 
-  @Override
-  protected KuduClient getKuduConnection(String masters) {
-    return mock(KuduClient.class);
-  }
+    @Override
+    protected KuduClient createClient(final String masters) {
+        final KuduClient client = mock(KuduClient.class);
 
-  @Override
-  protected KuduSession getKuduSession(KuduClient client){
-    return session;
-  }
+        try {
+            when(client.openTable(anyString())).thenReturn(mock(KuduTable.class));
+        } catch (final Exception e) {
 
-  @Override
-  protected KuduTable getKuduTable(KuduClient client, String tableName) throws KuduException {
-    return mock(KuduTable.class);
-  }
+        }
+
+        return client;
+    }
+
+    @Override
+    protected KuduSession getKuduSession(KuduClient client) {
+        return session;
+    }
 }
