+++ nifi-nar-bundles/nifi-kudu-bundle/nifi-kudu-processors/src/test/java/org/apache/nifi/processors/kudu/TestPutKudu.java
@@ -17,7 +17,6 @@
 
 package org.apache.nifi.processors.kudu;
 
-import java.util.stream.Collectors;
 import org.apache.kudu.ColumnSchema.ColumnSchemaBuilder;
 import org.apache.kudu.ColumnTypeAttributes;
 import org.apache.kudu.Schema;
@@ -41,14 +40,13 @@ import org.apache.nifi.serialization.RecordReader;
 import org.apache.nifi.serialization.RecordReaderFactory;
 import org.apache.nifi.serialization.SimpleRecordSchema;
 import org.apache.nifi.serialization.record.MapRecord;
+import org.apache.nifi.serialization.record.MockRecordParser;
 import org.apache.nifi.serialization.record.RecordField;
 import org.apache.nifi.serialization.record.RecordFieldType;
-import org.apache.nifi.serialization.record.MockRecordParser;
 import org.apache.nifi.serialization.record.RecordSchema;
 import org.apache.nifi.util.MockFlowFile;
 import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
-
 import org.apache.nifi.util.Tuple;
 import org.junit.After;
 import org.junit.Assert;
@@ -59,22 +57,22 @@ import org.mockito.stubbing.OngoingStubbing;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Arrays;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.EXCEPTION;
+import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.FAIL;
+import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.OK;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-
-import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.OK;
-import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.FAIL;
-import static org.apache.nifi.processors.kudu.TestPutKudu.ResultCode.EXCEPTION;
 
 public class TestPutKudu {
 
@@ -246,23 +244,6 @@ public class TestPutKudu {
         testRunner.assertAllFlowFilesTransferred(PutKudu.REL_FAILURE, 1);
     }
 
-    @Test
-    public void testSkipHeadLineTrue() throws InitializationException, IOException {
-        createRecordReader(100);
-        testRunner.setProperty(PutKudu.SKIP_HEAD_LINE, "true");
-
-        final String filename = "testSkipHeadLineTrue-" + System.currentTimeMillis();
-
-        final Map<String,String> flowFileAttributes = new HashMap<>();
-        flowFileAttributes.put(CoreAttributes.FILENAME.key(), filename);
-
-        testRunner.enqueue("trigger", flowFileAttributes);
-        testRunner.run();
-        testRunner.assertAllFlowFilesTransferred(PutKudu.REL_SUCCESS, 1);
-
-        MockFlowFile flowFiles = testRunner.getFlowFilesForRelationship(PutKudu.REL_SUCCESS).get(0);
-        flowFiles.assertAttributeEquals(PutKudu.RECORD_COUNT_ATTR, "99");
-    }
 
     @Test
     public void testInsertManyFlowFiles() throws Exception {
